:bug::recycle:Bug Fixes and Feature Improvements
Details:
- Bug fix related to issue #4
- Improved test code
- Upgraded GitHub Actions version
- Updated dependencies
