+++ .github/workflows/test_githubpages_deploy_multi.yml
@@ -12,7 +12,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest]
-        python-version: ["3.11"]
+        python-version: ["3.12"]
     runs-on: ${{ matrix.os }}
     env:
       TZ: "Asia/Tokyo"
@@ -23,16 +23,16 @@ jobs:
         id: is_main
         shell: bash
         run: |
-          if [ "${{ github.ref }}" = "refs/heads/main" ] && [ "${{ matrix.python-version }}" = "3.11" ];then
+          if [ "${{ github.ref }}" = "refs/heads/main" ] && [ "${{ matrix.python-version }}" = "3.12" ];then
             echo "flag=1" >> $GITHUB_OUTPUT
           else
             echo "flag=0" >> $GITHUB_OUTPUT
           fi
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
         with:
           persist-credentials: false
           fetch-depth: 0
-      - uses: actions/setup-python@v4
+      - uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
       - name: Install FFmpeg
@@ -96,7 +96,7 @@ jobs:
     strategy:
       matrix:
         os: [macos-12]
-        python-version: ['3.11']
+        python-version: ['3.12']
     runs-on: ${{ matrix.os }}
     needs: [test_and_deploy_ubuntu]
     env:
@@ -108,16 +108,16 @@ jobs:
         id: is_main
         shell: bash
         run: |
-          if [ "${{ github.ref }}" = "refs/heads/main" ] && [ "${{ matrix.python-version }}" = "3.11" ];then
+          if [ "${{ github.ref }}" = "refs/heads/main" ] && [ "${{ matrix.python-version }}" = "3.12" ];then
             echo "flag=1" >> $GITHUB_OUTPUT
           else
             echo "flag=0" >> $GITHUB_OUTPUT
           fi
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
         with:
           persist-credentials: false
           fetch-depth: 0
-      - uses: actions/setup-python@v4
+      - uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
       - name: Install FFmpeg
