+++ pyproject.toml
@@ -30,7 +30,7 @@ taskipy = "^1.11.0"
 pytest-cov = "^4.1.0"
 pytest-mock = "^3.10.0"
 pre-commit = "^3.3.2"
-ruff = "^0.0.275"
+ruff = "^0.6.4"
 pytest-xdist = "^3.5.0"
 pytest-html = "^4.1.1"
